use of NavLink from react-router-dom
